consolidate modules app gui cmd, update docu to reflect changes
